pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "universal-decryptor"
version = "1.0.0"
authors = [
  { name="Lucas", email="lucas@example.com" },
]
description = "A powerful, modular Python-based tool designed to decrypt a wide variety of ciphers."
readme = "readme.md"
requires-python = ">=3.13"
dependencies = [
  "colorama==0.4.6",
  "nltk==3.9.1",
  "pyenchant==3.2.2",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/Universal-Decryptor"
"Bug Tracker" = "https://github.com/yourusername/Universal-Decryptor/issues"

[project.scripts]
universal-decryptor = "decryptor:main"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
